requirements.txt
numpy
scipy
huffman